+++ src/HomePage.js
@@ -1,0 +1,5 @@
+import React from "react"
+
+export default function HomePage() {
+  return <div>HomePage</div>
+}
