+++ src/css/Building.css
@@ -1,0 +1,72 @@
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  color: #1a1a3f;
+}
+
+body {
+  font-family: Arial, sans-serif;
+  background-color: #f9fafb;
+}
+
+/* .App > div {
+      border: 1px black solid;
+  }
+  
+  .App > div > * {
+      border: 1px blue solid;
+  } */
+
+.App {
+  /* border: 3px green solid; */
+  display: flex;
+  flex-direction: column;
+  gap: 50px;
+  padding: 40px 60px;
+  height: 100vh;
+  max-height: 100vh;
+}
+
+.header {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 50px;
+  border-radius: 10px;
+  padding: 20px;
+  background-color: white;
+}
+
+.body1 {
+  display: grid;
+  grid-template-columns: 2fr 1fr;
+  gap: 50px;
+}
+
+.body2 {
+  height: 200px;
+  /* border: 1px black solid; */
+  border-radius: 10px;
+  padding: 20px;
+  background-color: white;
+}
+
+.logo,
+.userActions {
+  font-size: 36px;
+  /* display: flex;
+      justify-content: center;
+      align-items: center; */
+}
+
+/* Responsive adjustments 
+  @media (max-width: 768px) {
+      .header {
+          grid-template-columns: 1fr;
+      }
+  
+      .body1 {
+          grid-template-columns: 1fr;
+      }
+  }
+  */
